Use double precision (dvec3) for all tag calculations
- TagWrapper now stores tags as std::vector<glm::dvec3>
- Remove float casting during load/save/setPoints
- main.cpp: remove manual conversion in drawTagsOnSlice
- Full precision maintained throughout the pipeline

+++ new_register/include/TagWrapper.hpp
@@ -13,7 +13,7 @@ struct minc2_tags;
  * TagWrapper
  * ---------
  * Provides a simple RAII C++ interface around the C minc2_simple tag API.
- * It loads a .tag file and stores the tag points as glm::vec3 objects and
+ * It loads a .tag file and stores the tag points as glm::dvec3 objects and
  * optional string labels. The class manages the underlying `minc2_tags`
  * structure and frees it automatically.
  */
@@ -30,9 +30,9 @@ public:
     void load(const std::string& path); // throws std::runtime_error on failure
 
     /** Get the loaded points.
-     * @return Vector of glm::vec3 representing tag coordinates.
+     * @return Vector of glm::dvec3 representing tag coordinates.
      */
-    const std::vector<glm::vec3>& points() const { return points_; }
+    const std::vector<glm::dvec3>& points() const { return points_; }
 
     /** Get the loaded labels (may be empty strings if none).
      * @return Vector of label strings, same size as points().
@@ -81,7 +81,7 @@ private:
 
     minc2_tags* tags_;                // Raw C structure allocated by minc2_simple
     int n_volumes_;                   // Number of volumes stored in the tag file
-    std::vector<glm::vec3> points_;   // Tag coordinates (first volume)
+    std::vector<glm::dvec3> points_;   // Tag coordinates (first volume)
     std::vector<std::string> labels_; // Optional labels (may be empty strings)
 };
